requirements.txt
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.4
imutils==0.5.4
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
six==1.16.0
